requirements.txt
Flask==2.2.2
imutils==0.5.4
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
scikit_image==0.19.3
